float_cast

+++ RetinaFace/train.py
@@ -188,7 +188,7 @@ def train_net(args, ctx, pretrained, epoch, prefix, begin_epoch, end_epoch,
     #lr_factor = 0.1
     #lr = base_lr * (lr_factor ** (len(lr_epoch) - len(lr_epoch_diff)))
 
-    lr_epoch = [int(epoch) for epoch in lr_step.split(',')]
+    lr_epoch = [float(epoch) for epoch in lr_step.split(',')]
     lr_epoch_diff = [epoch - begin_epoch for epoch in lr_epoch if epoch > begin_epoch]
     lr_iters = [int(epoch * len(roidb) / input_batch_size) for epoch in lr_epoch_diff]
